Make UB benchmarks unsigned

+++ bench/multiply.c
@@ -4,14 +4,14 @@
 //
 // This benchmark tests the software multiply implemenation.
 
-int input1_multiply[100];
-int input2_multiply[100];
-int results_multiply[100];
+unsigned int input1_multiply[100];
+unsigned int input2_multiply[100];
+unsigned int results_multiply[100];
 
-void multiply_data_init(int nonce) {
-    int i;
-    int x = nonce;
-    int y = nonce;
+void multiply_data_init(unsigned int nonce) {
+    unsigned int i;
+    unsigned int x = nonce;
+    unsigned int y = nonce;
 
     for (i = 0; i < 100; i++) {
         x = (x * 97 + 17) % 10009;
@@ -21,9 +21,9 @@ void multiply_data_init(int nonce) {
     }
 }
 
-int multiply(int x, int y) {
-    int i;
-    int result = 0;
+unsigned int multiply(unsigned int x, unsigned int y) {
+    unsigned int i;
+    unsigned int result = 0;
 
     for (i = 0; i < 32; i++) {
         if ((x & 0x1) == 1)
@@ -36,27 +36,27 @@ int multiply(int x, int y) {
     return result;
 }
 
-int verify_multiply(int n, int* test) {
-    int i;
-    int result = 0;
+int verify_multiply(unsigned int n, unsigned int* test) {
+    unsigned int i;
+    unsigned int result = 0;
 
     for (i = 0; i < n; i++) {
-        int t0 = input1_multiply[i];
-        int t1 = input2_multiply[i];
-        int v = results_multiply[i];
+        unsigned int t0 = input1_multiply[i];
+        unsigned int t1 = input2_multiply[i];
+        unsigned int v = results_multiply[i];
         if (t0 * t1 != v)
             return 1;
 
         result += v;
     }
 
-    return result;
+    return (int)result;
 }
 
 int run_multiply(int dummy_0, int nonce) {
-    int i;
+    unsigned int i;
 
-    multiply_data_init(nonce);
+    multiply_data_init((unsigned int)nonce);
     for (i = 0; i < 100; i++) {
         results_multiply[i] = multiply(input1_multiply[i], input2_multiply[i]);
     }
